refactor: 키워드와 맞는 이름이 match되는 user filtering하는 로직 커스텀 훅으로 분리
